@@ -1,111 +1,118 @@
+using BelotWebApp.Areas.Identity.Data;
+using BelotWebApp.BelotClasses;
+using BelotWebApp.Data;
+using BelotWebApp.Services.EmailService;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
-using BelotWebApp.Data;
-using BelotWebApp.Areas.Identity.Data;
-using BelotWebApp.Services.EmailService;
-using BelotWebApp.BelotClasses;
 
-var builder = WebApplication.CreateBuilder(args);
-var connectionString = builder.Configuration.GetConnectionString("AuthDbContextConnection") ?? throw new InvalidOperationException("Connection string 'AuthDbContextConnection' not found.");
-
-builder.Services.AddDbContext<AuthDbContext>(options => options.UseSqlServer(connectionString));
-
-builder.Services.AddDefaultIdentity<ApplicationUser>(options => options.SignIn.RequireConfirmedAccount = false)
-        .AddRoles<IdentityRole>()
-        .AddEntityFrameworkStores<AuthDbContext>();
-
-// Add services to the container.
-builder.Services.AddControllersWithViews();
-builder.Services.AddRazorPages();
-builder.Services.AddSignalR();
-
-builder.Services.AddScoped<IEmailSender, EmailService>();
-
-var app = builder.Build();
-
-// Configure the HTTP request pipeline.
-if (!app.Environment.IsDevelopment())
+internal class Program
 {
-    app.UseExceptionHandler("/Home/Error");
-    // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
-    app.UseHsts();
-}
-
-app.UseHttpsRedirection();
-app.UseStaticFiles();
-
-app.UseRouting();
-
-// Ensure roles exist at startup
-using (var scope = app.Services.CreateScope())
-{
-    var roleManager = scope.ServiceProvider.GetRequiredService<RoleManager<IdentityRole>>();
-    await EnsureRolesExist(roleManager);
-}
-
-// Grant primary admin
-using (var scope = app.Services.CreateScope())
-{
-    var userManager = scope.ServiceProvider.GetRequiredService<UserManager<ApplicationUser>>();
-    var config = scope.ServiceProvider.GetRequiredService<IConfiguration>();
-
-    var username = config["PrimaryAdmin:Username"];
-    var email = config["PrimaryAdmin:Email"];
-    var password = config["PrimaryAdmin:Password"];
-
-    var user = await userManager.FindByEmailAsync(email);
-
-    if (user == null)
+    private static async Task Main(string[] args)
     {
-        user = new ApplicationUser
-        {
-            UserName = username,
-            Email = email,
-            EmailConfirmed = true
-        };
+        var builder = WebApplication.CreateBuilder(args);
+        var connectionString = builder.Configuration.GetConnectionString("AuthDbContextConnection") ?? throw new InvalidOperationException("Connection string 'AuthDbContextConnection' not found.");
 
-        var createResult = await userManager.CreateAsync(user, password);
+        builder.Services.AddDbContext<AuthDbContext>(options => options.UseSqlServer(connectionString));
 
-        if (!createResult.Succeeded)
+        builder.Services.AddDefaultIdentity<ApplicationUser>(options => options.SignIn.RequireConfirmedAccount = false)
+                .AddRoles<IdentityRole>()
+                .AddEntityFrameworkStores<AuthDbContext>();
+
+        // Add services to the container.
+        builder.Services.AddControllersWithViews();
+        builder.Services.AddRazorPages();
+        builder.Services.AddSignalR();
+
+        builder.Services.AddScoped<IEmailSender, EmailService>();
+
+        var app = builder.Build();
+
+        // Configure the HTTP request pipeline.
+        if (!app.Environment.IsDevelopment())
         {
-            var errors = string.Join(", ", createResult.Errors.Select(e => e.Description));
-            throw new Exception($"Failed to create admin user: {errors}");
+            app.UseExceptionHandler("/Home/Error");
+            // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+            app.UseHsts();
         }
-    }
 
-    if (!await userManager.IsInRoleAsync(user, "Admin"))
-    {
-        await userManager.AddToRoleAsync(user, "Admin");
-    }
-}
+        app.UseHttpsRedirection();
+        app.UseStaticFiles();
 
+        app.UseRouting();
 
-
-app.UseAuthentication(); ;
-
-app.UseAuthorization();
-
-app.UseEndpoints(endpoints =>
-{
-    endpoints.MapHub<BelotRoom>("/belotroom/{roomId}");
-});
-
-app.MapControllerRoute(
-    name: "default",
-    pattern: "{controller=Home}/{action=Index}/{id?}");
-app.MapRazorPages();
-
-app.Run();
-
-static async Task EnsureRolesExist(RoleManager<IdentityRole> roleManager)
-{
-    string[] roles = new[] { "Player", "Admin" };
-
-    foreach (var role in roles)
-    {
-        if (!await roleManager.RoleExistsAsync(role))
+        // Ensure roles exist at startup
+        using (var scope = app.Services.CreateScope())
         {
-            await roleManager.CreateAsync(new IdentityRole(role));
+            var roleManager = scope.ServiceProvider.GetRequiredService<RoleManager<IdentityRole>>();
+            await EnsureRolesExist(roleManager);
+        }
+
+        // Grant primary admin
+        using (var scope = app.Services.CreateScope())
+        {
+            var userManager = scope.ServiceProvider.GetRequiredService<UserManager<ApplicationUser>>();
+            var config = scope.ServiceProvider.GetRequiredService<IConfiguration>();
+
+            var username = config["PrimaryAdmin:Username"];
+            var email = config["PrimaryAdmin:Email"];
+            var password = config["PrimaryAdmin:Password"];
+
+            if (!string.IsNullOrEmpty(username) && !string.IsNullOrEmpty(email) && !string.IsNullOrEmpty(password))
+            {
+                var user = await userManager.FindByNameAsync(username);
+
+                if (user == null)
+                {
+                    user = new ApplicationUser
+                    {
+                        UserName = username,
+                        Email = email,
+                        EmailConfirmed = true
+                    };
+
+                    var createResult = await userManager.CreateAsync(user, password);
+
+                    if (!createResult.Succeeded)
+                    {
+                        var errors = string.Join(", ", createResult.Errors.Select(e => e.Description));
+                        throw new Exception($"Failed to create admin user: {errors}");
+                    }
+                }
+
+                if (!await userManager.IsInRoleAsync(user, "Admin"))
+                {
+                    await userManager.AddToRoleAsync(user, "Admin");
+                }
+            }
+        }
+
+        app.UseAuthentication(); ;
+
+        app.UseAuthorization();
+
+        app.UseEndpoints(endpoints =>
+        {
+            endpoints.MapHub<BelotRoom>("/belotroom/{roomId}");
+        });
+
+        app.MapControllerRoute(
+            name: "default",
+            pattern: "{controller=Home}/{action=Index}/{id?}");
+        app.MapRazorPages();
+
+        app.Run();
+    }
+
+    private static async Task EnsureRolesExist(RoleManager<IdentityRole> roleManager)
+    {
+        string[] roles = ["Player", "Admin"];
+
+        foreach (var role in roles)
+        {
+            if (!await roleManager.RoleExistsAsync(role))
+            {
+                await roleManager.CreateAsync(new IdentityRole(role));
+            }
         }
     }
 }
